+++ extras/integration_tests/tests/lib/database/database.dart
@@ -52,12 +52,12 @@ class Preferences {
 class PreferenceConverter extends NullAwareTypeConverter<Preferences, String> {
   const PreferenceConverter();
   @override
-  Preferences requireMapToDart(String fromDb) {
+  Preferences requireFromSql(String fromDb) {
     return Preferences.fromJson(json.decode(fromDb) as Map<String, dynamic>);
   }
 
   @override
-  String requireMapToSql(Preferences value) {
+  String requireToSql(Preferences value) {
     return json.encode(value.toJson());
   }
 }